Http server: Additional file support
The patch series change the file api implementation to use enum over
string.

It also introduce the following operation:
GET
GETFILESTATUS - return information on a file
LISTSTATUS -  return information on all files in a directory

PUT
MKDIRS - create a directory
RENAME - rename a file or a directory
COPY - copy a file

The simplest way of testing the changes is to use the Swagger-UI file
operation.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@cloudius-systems.com>