Http Server: Move exception to their own file
Exception handling in the httpserver is cover by http exception, and
general exception.

Http exception are mainly a result of bad input: a resource is not found
or arguments are wrong General exception results from some error in the
code.

The exception file holds a base class for the exception and the basic
derivation from it (not found, bad request and missing parameter).  All
the http server exception has a message and a status code that will be
used in the reply.

The exception handler will now use the message and status from the
exception in all httpserver exception.

Other exception will result a 500 (internal server error) reply.

Reviewed-by: Tomasz Grabiec <tgrabiec@gmail.com>
Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>