trace: make timed-trace recognition handle discontinuous streams properly
This change aims to reduce the likelihood of running into issue
merged stream may not be. If samples cross CPUs and the CPU of the
second sample has its earliest sample after the first sample, we
ignore that pair as the second sample has a high chance of belonging
to other pair.

This change does not solve the problem completly because there may be
a situation when sample which was recorded on another CPU was
discarded and the next sample will be again on the first CPU.

Signed-off-by: Tomasz Grabiec <tgrabiec@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>