vmxnet3: Drop memory allocation error handling on RX path
Packet discard logic on RX path is incorrect because
it doesn't take into account that incoming packet may be spread
over a few rings.

At the same time discard logic is called in case of newbuf() failure
only and our newbuf() never fails since we have no memory allocation
errors, so discard logic may be safely dropped and this is an
excellent fix for the original problem.

Signed-off-by: Dmitry Fleytman <dmitry@daynix.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>