sched: remove a bit of dead code.
Before commit 202b2cccc14a43d088002a80f7add51e3e4bc4ce, the scheduler
was responsible for saving the FPU state, so we needed to know whether
the scheduler itself uses it or not. Now that the FPU state is always
saved at interrupt time, we no longer care whether or not the scheduler
uses the FPU, so we can drop this flag.

Also drop the optional "pseudo-float" (integer-based floating point
operations) support from the scheduler. This never had any real
advantage over the actual floating point, and now that we save the
FPU state unconditionally, it makes even less sense to avoid floating
point.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@cloudius-systems.com>