java: make runjava.jar domain all-privileged
When application installed a SecurityManager, child protection domains
would have been declined the right to read files from their jar. This
was making apache derby fail with this:

java.util.MissingResourceException: Can't find bundle for base name org.apache.derby.loc.drda.messages, locale en
	at java.util.ResourceBundle.throwMissingResourceException(ResourceBundle.java:1499)
	at java.util.ResourceBundle.getBundleImpl(ResourceBundle.java:1322)
	at java.util.ResourceBundle.getBundle(ResourceBundle.java:795)
	at org.apache.derby.iapi.tools.i18n.LocalizedResource.setResource(LocalizedResource.java:184)
	at org.apache.derby.iapi.tools.i18n.LocalizedResource.getTextMessage(LocalizedResource.java:311)
	at org.apache.derby.iapi.tools.i18n.LocalizedResource.getTextMessage(LocalizedResource.java:273)
	at org.apache.derby.impl.drda.NetworkServerControlImpl.localizeMessage(NetworkServerControlImpl.java:3536)
	at org.apache.derby.impl.drda.NetworkServerControlImpl.localizeMessage(NetworkServerControlImpl.java:3492)
	at org.apache.derby.drda.NetworkServerControl.installSecurityManager(NetworkServerControl.java:705)

The bundle actually exists and resides in derbynet.jar but since we
have a SecurityManager installed and runjava.jar domain, which is a
parent of derbynet.jar, does not have permissions for reading other
jars than itself we fail on permission check inside the class
loader. When the same code is run on the host it passes the check
because derbynet.jar is invoked directly from an all-privileged
context. I think the solution is to make our runjava.jar
all-privileged by default. One way to do that is to grant the
permissions explicitly in the java.policy file, or put runjava.jar in
the extensions directory (/usr/lib/jvm/jre/lib/ext). The latter is
simpler so I went with that approach.