mmu: fix RCU page table walking
hw_ptep holds a pointer to a pt_element in rcu_ptr object, but this
is not the pointer that is protected by RCU, hw_ptep object are always
live on a stack and never accessed from more than one thread. The one
that is protected by RCU and can be changed concurrently resides in page
table itself. Fix that by making a pointer to point to atomic type and
use correct memory order during various accesses.

Signed-off-by: Gleb Natapov <gleb@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@cloudius-systems.com>