zfs: Port relatime property from ZFSonLinux
OSv port details:
* Adapted zfs_register_callbacks changes
* Discarded man page changes
* Solved conflict in zfs/zfs_znode.c
* Solved conflict in .../opensolaris/lib/libzfs/common/libzfs_dataset.c
* Added lacking MTNENT definitions into .../opensolaris/sys/mntent.h
* Adapted zfs_tstamp_update_setup and zfs_atime_need_update

Description:
"Add the "relatime" property.  When set to "on", a file's atime will only
be updated if the existing atime at least a day old or if the existing
ctime or mtime has been updated since the last access.  This behavior
is compatible with the Linux "relatime" mount option."

Reviewed-by: Glauber Costa <glommer@cloudius-systems.com>
Signed-off-by: Raphael S. Carvalho <raphaelsc@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>