sysconf: available number of physpages
While we correctly report the number of physical pages in the system, we
totally fail to report the number of available pages in the system - which is
used by Java's os.getFreePhysicalMemorySize().

This may lead to applications doing stupid decisions with, for instance, heap
vs off-heap placement. Example of Apache's solr log report:

Before:
sysconf(): stubbed for parameter 86
System Memory: 100.0001% used of 3.9996223 GiB

After:
System Memory: 2.9508517% used of 3.9996223 GiB

Unfortunately, that didn't help with the benchmark results as I was expecting
it could.  But it is wrong nevertheless.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>