Http Server: Move the parameters object to the request
The parameter map, holds path parameters. For example the if a url
matching rule is defined as /file/{path} and a url is /file/etc/hosts
the param object will hold an entry for "path" => "/etc/hosts".

To simplify the API the parameters object was moved to the request
object. It can only be calculated by the routes, as it depends on
specific handler (ie. /file/etc/{path} and /file/{path} will have
different value in the path entry)

For that the routes will get the request as non const reference and
will use the parameters object that is now in the request.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@cloudius-systems.com>