vfs: add support to follow symlink in namei
namei functions resolves path and returns it's directory entry. namei
processes each component of the path separately. To ensure correct
processing of the path, namei should follow any symbolic link it may
encounter. Symbolic link is a pointer to another directory entry. It
contains path to which it points. To follow a symbolic link, the path
to which it points must be read first. A standard system call to read
path to which symbolic link points is readlink.

A readlink reads value of symbolic link and copies it in return buffer.
However, for readlink to work, last component of given path must not
be followed.

In summary, this patch
1. Modifies namei to follow symbolic link
2. Adds a new API namei_last_nofollow() which works like namei, except
if last component of the path is symbolic link it won't follow it.
3. Adds readlink functionality

V1:
- changed return prototype of sys_readlink
- added lock in read_link before reading link value
- changed plan char * to std::unique_ptr
- refactored namei() code and removed a goto

Signed-off-by: Prasad Joshi <prasadjoshi.linux@gmail.com>
Signed-off-by: Avi Kivity <avi@cloudius-systems.com>