OSv Init: the main
The main starts an osvinit task. It should get in the command line
either a file name or a server name and path on the server as the init
file.

It uses a parser for passing the configuration to the osvinit

It call the osvinit wait, it would return imidiatialy if the httpserver
was not run.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@cloudius-systems.com>