OSv Init: add the module.py
Package the yaml-cpp lib, the osvinit.so and the osvinit.yaml init file.
By default the osvinit will not be started, but the httpserver will.

To change that, the starting line can be uncommentted or the osvinit can
be started with a -e directly.

The osvinit and httpserver should not be both started.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>

module.py need merge
Signed-off-by: Avi Kivity <avi@cloudius-systems.com>