vfs: bio: call bio_sync when closing the vfs layer
bio_sync is a function intended to be called when the bio layer is
done. It goes through all the cached buffers, and sync buffers that
were marked as B_DELWRI, meaning that I/O was delayed until the
underlying buffer is reused. Simply put, delayed writes.

If OSv is being powered off, it's unlikely that buffers marked as
B_DELWRI will be re-used in time. This patch is not fixing a bug as
ZFS currently doesn't make use of delayed writes nor does our vfs
layer. Anyway, bio_sync is the de-initialize function for the bio
layer, and should be called when closing the vfs layer.

Signed-off-by: Raphael S. Carvalho <raphaelsc@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@cloudius-systems.com>