elf: warn when using unsupported combination of PIE and TLS
As explained in issue #352, a PIE (position-independent executable) using
TLS (thread-local storage) does not currently work correctly. Fixing this
case is not easy, but until we do, it is dangerous to allow this case to
be ignored, as it causes reading and writing to wrong positions in memory
whenever TLS is used in the PIE.

So this patch adds a warning printout (to stdout) whenever one attempts to
load a PIE which has any TLS variables.

This is, of course, just a temporary measure. The real solution is to fix
object load and not just warn.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@cloudius-systems.com>