tcp: trim mbuf to IP packet length in the net channel path
Fixes #342.

According to RFC894 the minimum ethernet frame size is 60 bytes. When
the packet is smaller than that it should be padded. Looks like vmxnet
in constrast to virtio-net respects this rule and pads small packets
with zeros.

FreeBSD's stack is prepared for this, the mbuf is trimmed in
ip_preprocess_packet() which is called from ip_input(). However our
net channel fast path did not have this logic. This patch adds the
missing processing to tcp_net_channel_packet().

Reported-by: Lyor Goldstein <lgoldstein@vmware.com>
Signed-off-by: Tomasz Grabiec <tgrabiec@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@cloudius-systems.com>