tests: do not rely on osv::power_off() to sleep
Nadav saw tracing_smoke_test failing when run inside a VM on QEMU
without KVM. The test failed because there was no trace of
osv::power_off() in the backtrace of 'trace prof-wait'. We concluded
that it is best not to rely on this function to always sleep.

I changed the assertion to use 'sched::thread::wait' instead, which
is so generic it should probably always be there.

Reported-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Tomasz Grabiec <tgrabiec@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>