sched: Add API for iterating over all threads
We currently don't provide any method of iterating over all threads (something
which, "top" would like to do, for example). thread_map and thread_map_mutex
are private to sched.cc, and should stay that way.

So this patch provides a new function with_all_friends(f), which iterates
over all threads and calls the given std::function for each one. The
function is supposed to be safe (a thread will not be freed while someone
is running f on it), but it is inefficient as it holds thread_map_mutex
throughout the entire iteration, so it should currently only be used for
debugging purposes, e.g., for implementing "top".

Reviewed-by: Glauber Costa <glommer@cloudius-systems.com>
Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>