console: Fix echo on VGA console
This fixes #360, which is about input echo not being visible on the VGA
console until a newline is pressed - while it is immediately visible on
the serial console.

As part of the multi-console work (see commit c0a1e555), a "console_driver"
interface layer was added whose write side includes not just the write()
function we always had, but now also a flush(). In the serial console
implementation, the flush() does nothing, but in the VGA console
implementation, nothing is displayed until either a flush() or a newline
is written. So in line-discipline.cc, it is no longer enough to call
write() to echo the imput characters - we now need to also add a call
to flush() after every write(). Adding it solves this bug.

This does not solve the backspace not working on the VGA console. This is a
separate issue that will be handled in a separate patch.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>