kbd: change mapping of backspace from ^H to ^?
This patch makes backspace work correctly on the VGA console (this is
part of issue #360).

Our "multiplexed console" has one termios(3) setting, but two distinct
keyboards - the serial console input and the PS/2 keyboard. We defaulted
the VERASE character to ^? (\177, ASCII "DEL"), expecting that this is what
we will get on the serial console when a user presses the backspace key,
but our PS/2 keyboard driver generated a ^H (\10, ASCII "BS") and therefore
was not recognizing backspaces.

So this patch changes the PS/2 keyboard driver to generate a ^? on
backspace press, instead of ^H. This is what the Linux console, and most
Linux terminal emulators, also generate for this key.

I have to admit that I am really saddened by sending this patch. For
years I've been warning against the slow but sure attempts to displace
the ASCII BS character - which backspaces used to generate, and
use the DEL character instead. For years I've insisted on keeping
backspace==BS on my systems, and this is also how my home desktop is
set up. But I've obviously lost, and all modern Linux distributions
return ^?, not ^H, when a backspace is pressed, and set VERASE to ^?.

Some of my old posts about this backspace/delete issue:
http://copilotco.com/mail-archives/linux-kernel.2001/msg47697.html
https://www.mail-archive.com/linux-il@cs.huji.ac.il/msg52834.html

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>