Document dependency on python-requests
It is the state-of-the-art library for writing HTTP clients in
python. Its supoprt for SSL is much better than that of urllib or
urllib2. The latter do not provide an easy way to do certificate-based
client authentication which is an immediate reason for switching
to it.

The package is readily available in both Fedora and Ubuntu
distributions.

See: http://docs.python-requests.org/

This package will be used by python clients like trace.py and top.py