command line: don't wait for disowned threads
If we run

        run.py -e "command1 & command2"

This runs command1 and command2 in parallel. Only when the two commands
return from main(), OSv will shut down.

But for some service background threads, delaying the shutdown is not
a feature, it's a bug. For example if you run,

        run.py -e "libhttpserver.so & /tests/tst-chdir.so"

the guest will never exit, despite tst-chdir.so finishing in less than a
second, because libhttserver never completes.

It's even more annoying now that we automatically add "libhttpserver.so &"
to every command line - now even a 'run.py -e /tests/tst-chdir.so' will
never complete.

This patch adds a new alternative to "&": "&!". This syntax is inspired
by zsh, where &! means creating a *disowned* background process - one which
the shell doesn't track, and doesn't worry about when the shell exits.

With this patch,

        run.py -e "libhttpserver.so &! /tests/tst-chdir.so"

Will shutdown as soon as chdir.so finishes - no matter if libhttpserver.so
finished (it never will...).

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>