loader: always run httpserver, and other stuff from /init
Some modules, especially httpserver, have a command line which we *always*
want to run, regardless of run.py's "-e" paramter. For example, if the
user builds an image with "make image=httpserver,memcached", the default
is to run both, but if someone uses run.py -e to run memcached with
special parameters, httpserver is not run.

This patch fixes this annoyance by adding a new mechanism to run a
command line for each module on boot time:

The idea is that files in the directory "/init/" specify command lines that
will be run, at filename alphabetical order, before the "-e" command line.
The contents of each /init/* file is our usual command line syntax
(supporting quotes, ";", "&" and "&!").

With this patch, the "httpserver" module no longer has a classic command-line
setting. Rather, it includes a file /init/10-httpserver, with the contents

        /libhttpserver.so &!

This command line will *always* be run on boot, regardless of whether or
not a command line is specified with run.py's "-e" option. The "&!" means
that if the main command line finishes, OSv will shut down, without waiting
for libhttpserver.so to finish (that would never happen).

A "--noinit" kernel option is provided to override the /init/* files
and not run them.

Fixes #358.

Reviewed-by: Tomasz Grabiec <tgrabiec@gmail.com>
Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>