random: create random devices even when there's no virtio-rng
This fixes a bug which manifests in /dev/random and /dev/urandom
missing on platforms which do not have virtio-rng. That includes GCE
and EC2. The problem was that if virtio-rng did not register itself
then random devices were not initialized. We now have more hardware
sources (rdrand) and we should not rely on virito-rng alone.

The fix is to always initialize random devices. If virtio-rng is
probed successfully it will register itself as entropy source.

Reviewed-by: Raphael S. Carvalho <raphaelsc@cloudius-systems.com>
Signed-off-by: Tomasz Grabiec <tgrabiec@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>