Http Server Testing: Adding test for thread API
This test check that the thread API timestamp progress and that the
thread idle time cpu_ms counter progress.

The test itself calls the thread API twice with 2s time difference,
it look for the idle thread and check that it has the same id and that
its cpu_ms was changed.

Reviewed-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>