Add missing backtrace line
As explained in issue #380, too often our backtrace results are bogus.

One of the problems was that very often the function itself in which the
crash happened was missing from the backtrace. It turns out that
print_backtrace() function always assumed it should ignore two frames,
for abort() which calls abort("Aborted."). However, when abort(message)
was called directly, and this was often the case, we unnecessarily
ignored one interesting frame.

So in this patch we ignore just the first frame (abort(const char*)),
and later ignore an abort() frame, but only if there is one.

Note that there are still bogus backtrace problems - this is only a
partial solution.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>